design: 카메라, 캡션뷰 디자인 디테일 반영
